# pitsa/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pitsa",
    "description": "PiTSA API",
    "license": {
      "name": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/api/v1/tsp": {
      "post": {
        "tags": [
          "tsp_resources"
        ],
        "summary": "Time-Stamp Protocol via HTTP as defined in\n[RFC3161 3.4](https://www.rfc-editor.org/rfc/rfc3161#section-3.4).",
        "operationId": "tsp_raw_time_stamp_request",
        "requestBody": {
          "description": "DER encoded TimeStampReq.",
          "content": {
            "application/timestamp-query": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok. Parse the DER encoded TimeStampResponse for actual status defined in the time-stamping protocol.",
            "content": {
              "application/timestamp-reply": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Bad Reqest"
          }
        }
      }
    },
...
  },
...
}